--- token-issuer-api/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = 'token-issuer-api'
