--- token-issuer-api/src/main/java/vn/chuot96/jwtissapi/compo/ExternalFileReader.java
@@ -1,48 +1,0 @@
-package vn.chuot96.jwtissapi.compo;
-
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Paths;
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.stream.Stream;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
-
-@Component
-public class ExternalFileReader {
-    private final Map<String, String> data = new HashMap<>();
-
-    @Value("${external-file.path}")
-    private String path;
-
-    private boolean loaded = false;
-
-    private void loadFileOnce() {
-        if (!loaded) {
-            try (Stream<String> stream = Files.lines(Paths.get(path))) {
-                stream.forEach(line -> {
-                    String[] parts = line.split("=", 2);
-                    if (parts.length == 2) {
-                        data.put(parts[0].trim(), parts[1].trim());
-                    }
-                });
-                loaded = true;
-            } catch (IOException e) {
-                throw new RuntimeException("Cannot read external file: " + path, e);
-            }
-        }
-    }
-
-    public String string(String key) {
-        loadFileOnce();
-        return data.get(key);
-    }
-
-    public List<String> list(String key) {
-        String raw = string(key);
-        return (raw != null) ? Arrays.stream(raw.split(",")).map(String::trim).toList() : List.of();
-    }
-}
