--- token-issuer-api/src/main/java/vn/chuot96/jwtissapi/dto/RequestDTO.java
@@ -1,5 +1,0 @@
-package vn.chuot96.jwtissapi.dto;
-
-import java.util.List;
-
-public record RequestDTO(String provider, String subject, String scope, List<String> audience) {}
