--- token-issuer-api/src/main/java/vn/chuot96/jwtissapi/dto/ResponseDTO.java
@@ -1,3 +1,0 @@
-package vn.chuot96.jwtissapi.dto;
-
-public record ResponseDTO(String access, String refresh) {}
