--- token-issuer-api/src/main/java/vn/chuot96/jwtissapi/service/TokenService.java
@@ -1,20 +1,0 @@
-package vn.chuot96.jwtissapi.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import vn.chuot96.jwtissapi.compo.TokenCustomize;
-import vn.chuot96.jwtissapi.dto.RequestDTO;
-
-@Service
-@RequiredArgsConstructor
-public class TokenService {
-    private final TokenCustomize tokenCustomize;
-
-    public String issueAccess(RequestDTO request) {
-        return tokenCustomize.generateAccess(request);
-    }
-
-    public String issueRefresh(RequestDTO request) {
-        return tokenCustomize.generateRefresh(request);
-    }
-}
