--- token-issuer-api/src/main/resources/application-issuer1.properties
@@ -1,13 +1,0 @@
-spring.application.name = token-issuer-api-1
-#___________________________________________
-server.port = 8101
-#___________________________________________
-jwt.issuer = http://${spring.application.name}:${server.port}
-#___________________________________________
-# ======= Eureka Server Registration =======
-eureka.client.service-url.defaultZone = http://localhost:8761/eureka
-eureka.instance.prefer-ip-address = true
-eureka.instance.instance-id = ${spring.application.name}:${server.port}
-#___________________________________________
-# =========== External File Path ===========
-external-file.path = E:/Document/texts/data.txt
