--- token-issuer-api/src/test/java/vn/chuot96/jwtissapi/TokenDurationIssuerAPITests.java
@@ -1,11 +1,0 @@
-package vn.chuot96.jwtissapi;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class TokenDurationIssuerAPITests {
-
-    @Test
-    void contextLoads() {}
-}
