+++ token-issuer/src/main/java/vn/chuot96/jwtiss/config/SecurityConfig.java
@@ -1,0 +1,29 @@
+package vn.chuot96.jwtiss.config;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
+import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+import vn.chuot96.jwtiss.compo.HeaderValueFilter;
+
+@Configuration
+@RequiredArgsConstructor
+public class SecurityConfig {
+    private final HeaderValueFilter headerValueFilter;
+
+    @Bean
+    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
+        http.csrf(AbstractHttpConfigurer::disable)
+                .authorizeHttpRequests(auth -> auth.requestMatchers("/issue/**")
+                        .authenticated()
+                        .anyRequest()
+                        .permitAll())
+                .addFilterBefore(headerValueFilter, UsernamePasswordAuthenticationFilter.class)
+                .sessionManagement(session -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS));
+        return http.build();
+    }
+}
