+++ token-issuer/src/main/java/vn/chuot96/jwtiss/dto/RequestDTO.java
@@ -1,0 +1,5 @@
+package vn.chuot96.jwtiss.dto;
+
+import java.util.List;
+
+public record RequestDTO(String provider, String subject, String scope, List<String> audience) {}
