+++ token-issuer/src/main/java/vn/chuot96/jwtiss/dto/ResponseDTO.java
@@ -1,0 +1,3 @@
+package vn.chuot96.jwtiss.dto;
+
+public record ResponseDTO(String access, String refresh) {}
