+++ token-issuer/src/main/resources/application-issuer1.properties
@@ -1,0 +1,14 @@
+application.code=tbirpg2d
+spring.application.name = token-issuer-${application.code}
+#___________________________________________
+server.port = 8101
+#___________________________________________
+jwt.issuer = http://${spring.application.name}:${server.port}
+#___________________________________________
+# ======= Eureka Server Registration =======
+eureka.client.service-url.defaultZone = http://localhost:8761/eureka
+eureka.instance.prefer-ip-address = true
+eureka.instance.instance-id = ${spring.application.name}:${server.port}
+#___________________________________________
+# =========== External File Path ===========
+external-file.path = E:/Document/texts/data.txt
