+++ token-issuer/src/test/java/vn/chuot96/jwtiss/TokenIssuerTests.java
@@ -1,0 +1,11 @@
+package vn.chuot96.jwtiss;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class TokenIssuerTests {
+
+    @Test
+    void contextLoads() {}
+}
